+++ MOP/src/Systems/ExceptionManager.cs
@@ -104,11 +104,12 @@ namespace MOP
             output += $"ActiveDistanceMultiplicationValue: {MopSettings.ActiveDistanceMultiplicationValue}\n";
             output += $"SafeMode: {MopSettings.SafeMode}\n";
             output += $"RemoveEmptyBeerBottles: {MopSettings.RemoveEmptyBeerBottles}\n";
-            output += $"SatsumaTogglePhysicsOnly: {MopSettings.SatsumaTogglePhysicsOnly}\n";
             output += $"ToggleVehiclePhysicsOnly: {Rules.instance.SpecialRules.ToggleAllVehiclesPhysicsOnly}\n";
             output += $"IgnoreModVehicles: {Rules.instance.SpecialRules.IgnoreModVehicles}\n";
             output += $"EnableFramerateLimiter: {((bool)MOP.EnableFramerateLimiter.GetValue())}\n";
             output += $"FramerateLimiter: {MOP.FramerateLimiter.GetValue()}\n";
+            output += $"EnableShadowAdjusting: {((bool)MOP.EnableShadowAdjusting.GetValue())}\n";
+            output += $"ShadowDistance: {MOP.ShadowDistance.GetValue()}\n";
             output += $"RulesAutoUpdate: {MOP.RulesAutoUpdate.GetValue()}\n"; 
             output += $"RulesAutoUpdateFrequency: {MopSettings.GetRuleFilesUpdateDaysFrequency()}\n";
             output += $"CustomRuleFile: {File.Exists($"{MOP.ModConfigPath}/Custom.txt")}\n\n";